+++ docs/changelog.md
@@ -1,6 +1,6 @@
 ﻿# Changelog
 
-## Upcoming version
+## 1.0.0-beta.2
 
 ### Common changes